refactor: SignUpRouteTest의 Snackbar 테스트 추가
